Version Packages (#711)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.

# Releases
## @hashicorp/react-subnav@9.4.0

### Minor Changes

-   [#705](https://github.com/hashicorp/react-components/pull/705) [`a361fab5`](https://github.com/hashicorp/react-components/commit/a361fab5f38ced326a1e291dc370559eaced3d6d) Thanks [@alexcarpenter](https://github.com/alexcarpenter)! - Adds menuText prop