Version Packages (#721)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.

# Releases
## @hashicorp/react-error-view@0.1.0

### Minor Changes

-   [#719](https://github.com/hashicorp/react-components/pull/719) [`883764e3`](https://github.com/hashicorp/react-components/commit/883764e37b9ce009a109efe0c14d82d7896e8691) Thanks [@nandereck](https://github.com/nandereck)! - Replaces track event name with static value.