Version Packages (#751)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.

# Releases
## @hashicorp/react-card@0.3.0

### Minor Changes

-   [#748](https://github.com/hashicorp/react-components/pull/748) [`be14f118`](https://github.com/hashicorp/react-components/commit/be14f118a5fc62b0abbee8d8ba23b5cb3c87e38e) Thanks [@nandereck](https://github.com/nandereck)! - Add ProductBadge support to Card