Version Packages (#753)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.

# Releases
## @hashicorp/react-related-content@0.2.0

### Minor Changes

-   [#731](https://github.com/hashicorp/react-components/pull/731) [`2a26b0bf`](https://github.com/hashicorp/react-components/commit/2a26b0bf09bc94fbee4607ae7a2b321bfb99a829) Thanks [@nandereck](https://github.com/nandereck)! - Create RelatedContent component