Make tile `background-color` white to support different background colors (#759)
🎟️ [Asana Task](https://app.asana.com/0/1202802428287999/1202891942449714/f)
🔍 [Preview Link](https://react-components-git-{branch-slug}-hashicorp.vercel.app)

---

<!-- Reminder: This is an open source project, make sure not to include any sensitive information in the pull request. -->

## Description
As part of Use Case QA, design requested that the `NextSteps` component support a gray background. While it's unclear at the moment if we want to add this functionality to the `NextSteps` component more broadly, we do want to be able to do this specifically for Use Cases. This is easily achieved by wrapping `NextSteps` in a div and setting the `background-color`, however, it was discovered that the non-primary tiles have a transparent background. This results in the tile and `NextSteps` components being the same color which is not desireable.

This PR sets the `background-color` for non-primary tiles to white so that changes to the wrapping element's `background-color` don't also change the `background-color` of the tile.

## Validation Steps
- Navigate to the [`Next Steps` component in the preview](https://react-components-git-nels-next-stepsadd-backgr-0355cd-hashicorp.vercel.app/components/nextsteps).
- Looking at the light and dark themes, validate that there are no visual regressions

### PR Checklist 🚀

Items in this checklist may not may not apply to your PR, but please consider each item carefully.

- [ ] Add Asana and Preview links above.
- [ ] Conduct thorough self-review.
- [ ] Add or update tests as appropriate.
- [ ] Conduct reasonable cross browser testing for both compatibility and responsive behavior (We have a [Sauce Labs](https://app.saucelabs.com/) account for this, if you don't have access, just ask!).
- [ ] Conduct reasonable accessibility review (use the [WAS](https://accessible.org/Web-Accessibility-Standards-WAS-2.pdf) as a guide or an [axe browser plugin](https://www.deque.com/axe/) until we establish more formal checks).
- [ ] Identify (in the description above) and document (add Asana tasks on [this board](https://app.asana.com/0/1100423001970639/list)) any technical debt that you're aware of, but are not addressing as part of this PR.