Version Packages (#763)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.

# Releases
## @hashicorp/react-next-steps@0.2.0

### Minor Changes

-   [#759](https://github.com/hashicorp/react-components/pull/759) [`4957b882`](https://github.com/hashicorp/react-components/commit/4957b882d2a6321d3a9921507c7dff4b700e1198) Thanks [@nandereck](https://github.com/nandereck)! - Set non-primary tile background color to white