Version Packages (#838)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.

# Releases
## @hashicorp/react-inline-video@0.3.0

### Minor Changes

-   [#837](https://github.com/hashicorp/react-components/pull/837) [`ce8ed5a4`](https://github.com/hashicorp/react-components/commit/ce8ed5a4725f9fff4638a5f3528bd1b559517d58) Thanks [@nandereck](https://github.com/nandereck)! - Add lazy loading to prevent hydration errors

 ## @hashicorp/react-video-feature@0.1.2

### Patch Changes

-   Updated dependencies \[[`ce8ed5a4`](https://github.com/hashicorp/react-components/commit/ce8ed5a4725f9fff4638a5f3528bd1b559517d58)]:
    -   @hashicorp/react-inline-video@0.3.0