Version Packages (#847)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.

# Releases
## @hashicorp/react-related-content@0.3.0

### Minor Changes

-   [#845](https://github.com/hashicorp/react-components/pull/845) [`4734c654`](https://github.com/hashicorp/react-components/commit/4734c65464f80bcfa976be24639b8fd66bb20c6c) Thanks [@nandereck](https://github.com/nandereck)! - Replace StandaloneLink with custom link