Version Packages (#863)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.

# Releases
## @hashicorp/react-badge@0.2.0

### Minor Changes

-   [#862](https://github.com/hashicorp/react-components/pull/862) [`23ba7a28`](https://github.com/hashicorp/react-components/commit/23ba7a2814fc0ce01da6a519e6554a8d8eadf610) Thanks [@nandereck](https://github.com/nandereck)! - Refactor to match WPL spec