Version Packages (#871)
Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>

This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.

# Releases
## @hashicorp/react-next-steps@0.8.0

### Minor Changes

-   [#866](https://github.com/hashicorp/react-components/pull/866) [`16f0acbc`](https://github.com/hashicorp/react-components/commit/16f0acbc1bf7e25ba3bf9b0b9af30e5e5aa60b1b) Thanks [@nandereck](https://github.com/nandereck)! - enable tertiary CTA, change tile colors and enable badges