fix(placeholder): Shine height
This PR fixes an issue with `.box` `background-size` causing the "shine" animation not to fill the box height on Safari/iOS.

Context: `.box` had no explicit height value set for its `background-size` property, and values default to auto if not set. However, in Safari/iOS, the `auto` for the height value is not the height of the container, but (from observation) the value set for width. This behavior seems to have caused the "shine" to only reach `75%` of the box's height in Safari/iOS.