Version Packages (#986)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.

# Releases
## @hashicorp/react-consent-manager@9.2.0

### Minor Changes

-   [#981](https://github.com/hashicorp/react-components/pull/981) [`0aedf445`](https://github.com/hashicorp/react-components/commit/0aedf44529c1a07bde59b9cc196db06635dcb305) Thanks [@nandereck](https://github.com/nandereck)! - Add onload functionality to scripts handled by ConsentManager